requirements.txt
streamlit>=1.37
langchain>=0.2
langchain-community>=0.2
langchain-groq>=0.1
langgraph>=0.2
chromadb>=0.5
sentence-transformers>=3.0
pypdf>=4.2
beautifulsoup4>=4.12
requests>=2.32
python-dotenv>=1.0
tiktoken>=0.7